+++ src/components/Logo/index.js
