+++ src/components/TypingEffect/index.js
